adjusted podmigration for debugging purposes

+++ src/main/java/com/example/monitoringservice/PodNodeAffinityHandler.java
@@ -75,14 +75,16 @@ public class PodNodeAffinityHandler {
 //                pod.getSpec().affinity(affinity);
 
 
-//                pod.getSpec().setNodeName(destinedNode.getMetadata().getName());
-                pod.getSpec().putNodeSelectorItem("kubernetes.io/hostname", destinedNode.getMetadata().getName());
+                if (pod.getSpec().getNodeName() != null && destinedNode.getMetadata().getName() != null && pod.getSpec().getNodeName().equals(destinedNode.getMetadata().getName())) {
+//                    pod.getSpec().putNodeSelectorItem("kubernetes.io/hostname", destinedNode.getMetadata().getName());
+                    pod.getSpec().setNodeName(destinedNode.getMetadata().getName());
 
-                System.out.println("Moving Pod: " + pod.getMetadata().getName() + "to Node: " + destinedNode.getMetadata().getName());
+                    System.out.println("Moving Pod: " + pod.getMetadata().getName() + "to Node: " + destinedNode.getMetadata().getName());
 //
-                api.deleteNamespacedPod(Objects.requireNonNull(pod.getMetadata()).getName(), "default", null, null, 0, null, "Background", null);
+                    api.deleteNamespacedPod(Objects.requireNonNull(pod.getMetadata()).getName(), "default", null, null, 0, null, "Background", null);
 
-                api.createNamespacedPod("default", pod, null, null, null, null);
+                    api.createNamespacedPod("default", pod, null, null, null, null);
+                }
 
 
 //                replacePodOnceTerminated(pod);
